+++ test/test_dna.py
@@ -1090,6 +1090,76 @@ Mismatches: 6/16 (37.50%)
             matchToString(match, read1, read2),
         )
 
+    def testAmbiguousMatchLocations(self):
+        """
+        Ambiguous match locations must be returned correctly.
+        """
+        read1 = Read('id1', 'TWTTTAAAAAAGCGCG')
+        read2 = Read('id2', 'TTTTT------GCGCK')
+        match = compareDNAReads(read1, read2)
+        self.assertEqual(
+            '''\
+Exact matches: 8/16 (50.00%)
+Ambiguous matches: 2/16 (12.50%)
+Exact or ambiguous matches: 10/16 (62.50%)
+Mismatches: 6/16 (37.50%)
+  Not involving gaps (i.e., conflicts): 0
+  Involving a gap in one sequence: 6/16 (37.50%)
+  Involving a gap in both sequences: 0
+  Involving no coverage in one sequence: 0
+  Involving no coverage in both sequences: 0
+  Id: id1
+    Length: 16
+    Gaps: 0
+    No coverage: 0
+    Ambiguous: 1/16 (6.25%)
+  Id: id2
+    Length: 16
+    Gaps: 6/16 (37.50%)
+    Gap locations (1-based): 6, 7, 8, 9, 10, 11
+    No coverage: 0
+    Ambiguous: 1/16 (6.25%)
+Ambiguous matches:
+    2 W T
+    16 G K''',
+            matchToString(match, read1, read2, includeAmbiguousMatches=True)
+        )
+
+    def testNonGapMismatchLocations(self):
+        """
+        Non-gap mismatch locations must be returned correctly.
+        """
+        read1 = Read('id1', 'TATTTAAAAAAGCGCG')
+        read2 = Read('id2', 'TTTTT------GCGCC')
+        match = compareDNAReads(read1, read2)
+        self.maxDiff = None
+        self.assertEqual(
+            '''\
+Exact matches: 8/16 (50.00%)
+Ambiguous matches: 0
+Mismatches: 8/16 (50.00%)
+  Not involving gaps (i.e., conflicts): 2/16 (12.50%)
+  Involving a gap in one sequence: 6/16 (37.50%)
+  Involving a gap in both sequences: 0
+  Involving no coverage in one sequence: 0
+  Involving no coverage in both sequences: 0
+  Id: id1
+    Length: 16
+    Gaps: 0
+    No coverage: 0
+    Ambiguous: 0
+  Id: id2
+    Length: 16
+    Gaps: 6/16 (37.50%)
+    Gap locations (1-based): 6, 7, 8, 9, 10, 11
+    No coverage: 0
+    Ambiguous: 0
+Non-gap mismatches:
+    2 A T
+    16 G C''',
+            matchToString(match, read1, read2, includeNonGapMismatches=True)
+        )
+
     def testNoCoverageLocations(self):
         """
         No coverage locations must be returned correctly.